updated chat room to automatically put your username if the name is set to nothing
