+++ client/src/components/Chatroom/index.js
@@ -1,15 +1,23 @@
-import React, { useState } from "react";
+import React, { useContext, useState } from "react";
 import { io } from "socket.io-client";
+import globalContext from "../../context/globalContext";
 import Chat from "./Chat";
 import "./style.css";
 const socket = io.connect("localhost:3000");
 
 function Chatroom() {
+  const { user } = useContext(globalContext)
+  console.log(user);
   const room = "levelup";
   const [username, setUsername] = useState("");
   const [showChat, setShowChat] = useState(false);
   const joinChat = () => {
     if (username !== "") {
+      socket.emit("join_room", room);
+      setShowChat(true);
+    }
+    if(username === ""){
+      setUsername(user.userName)
       socket.emit("join_room", room);
       setShowChat(true);
     }
